% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGENS_MAIN.R
\name{calgsea}
\alias{calgsea}
\title{Calcualte GSEA enrichment}
\usage{
calgsea(dir_ti_gene, j, rank_dd, pval_cutoff)
}
\arguments{
\item{dir_ti_gene}{Data.frame includes TIGs, avg_lo2FC, regulated direction}

\item{j}{The tested sub-population}

\item{rank_dd}{Ranked disease DEGs}

\item{pval_cutoff}{pvalue threshold for gsea enrichment results}
}
\value{
GSEA enrichment results
}
\description{
Calcualte GSEA enrichment
}
