% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGENS_MAIN.R
\name{DisAssoc}
\alias{DisAssoc}
\title{Check the association between specific disease and monocytes sub-populations}
\usage{
DisAssoc(
  disease_data,
  reg_dir,
  cluster_id,
  cutoff,
  sub_type,
  ti_gene,
  pval_cutoff
)
}
\arguments{
\item{disease_data}{The input data.frame with colnames "gene", "value", ("subtype")}

\item{reg_dir}{Value should be "up" or "down"}

\item{cluster_id}{Numeric variable or vector}

\item{cutoff}{Select genes to gsea}

\item{sub_type}{which disease subtype should be tested}

\item{ti_gene}{reference database downloaded from github}

\item{pval_cutoff}{pvalue threshold for gsea enrichment results}
}
\value{
Enriched GSEA results for plotting or other downstream analysis
}
\description{
This tool will use GASE embedded in clusterProfiler to check the association between a specific
disease and identified monocytes sub-populations. For each input dataset, diseases could be
classified into sub-types. For each disease/sub-type, genes will be ranked based on lof2FC.
4 sub-populations would be calculated one by one.
}
