% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGENS_MAIN.R
\name{enriched_gene_to_pathway}
\alias{enriched_gene_to_pathway}
\title{export enriched genes for downstream analysis}
\usage{
enriched_gene_to_pathway(gsea_results)
}
\arguments{
\item{gsea_results}{The results from GSEA enrichment}
}
\description{
export enriched genes for downstream analysis
}
