% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea.R
\name{GSEAplot}
\alias{GSEAplot}
\title{GSEAplot}
\usage{
GSEAplot(
  x,
  geneSetID,
  sub_type,
  disease_name,
  cluster_id,
  reg_dir,
  title = "Enrichment Testing for ",
  color = "green",
  base_size = 11,
  rel_heights = c(1.5, 0.5, 1),
  subplots = 1:3,
  pvalue_table = FALSE,
  ES_geom = "line"
)
}
\arguments{
\item{x}{gseaResult object}

\item{geneSetID}{gene set ID}

\item{sub_type}{tested sub_type of disease}

\item{disease_name}{tested disease}

\item{cluster_id}{trained immunity sub-population}

\item{reg_dir}{regulated direction of TIGs}

\item{title}{plot title}

\item{color}{color of running enrichment score line}

\item{base_size}{base font size}

\item{rel_heights}{relative heights of subplots}

\item{subplots}{which subplots to be displayed}

\item{pvalue_table}{whether add pvalue table}

\item{ES_geom}{geom for plotting running enrichment score,
one of 'line' or 'dot'}
}
\value{
plot
}
\description{
GSEA plot that mimic the plot generated by broad institute's GSEA software
}
