% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea.R
\name{GSEAplot_Basic}
\alias{GSEAplot_Basic}
\title{plot_gsea.R is based on enrichplot: https://github.com/YuLab-SMU/enrichplot
devloped by Guangchuang YU}
\usage{
GSEAplot_Basic(
  x,
  geneSetID,
  disease_name,
  sub_type,
  cluster_id,
  reg_dir,
  by = "all",
  title = "Enrichment Testing for ",
  color = "black",
  color.line = "green",
  color.vline = "#FA5860",
  ...
)
}
\arguments{
\item{x}{GSEA result object}

\item{geneSetID}{same as geneSetID in GSEA}

\item{disease_name}{the tested disease}

\item{sub_type}{sub_type of tested disease}

\item{cluster_id}{trained immunity sub-population}

\item{reg_dir}{regulated direction of TIGs}

\item{color}{color of line segments}

\item{color.line}{color of running enrichment score line}

\item{color.vline}{color of vertical line which indicating the
maximum/minimal running enrichment score}
}
\value{
ggplot2 object
}
\description{
Plot GSEA result
}
