% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGENS_MAIN.R
\name{PlotAssoc}
\alias{PlotAssoc}
\title{Plot enrichment results}
\usage{
PlotAssoc(gsea_results)
}
\arguments{
\item{gsea_results}{The result from GSEA enrichment}
}
\value{
gseaplots
}
\description{
Plot enrichment results
}
