% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGENS_MAIN.R
\name{PreProDisData}
\alias{PreProDisData}
\title{Pre-Processing of input dataset}
\usage{
PreProDisData(disease_data, sub_type)
}
\arguments{
\item{disease_data}{data.frame with colnames "gene", "value", ("subtype")}

\item{sub_type}{which disease subtype should be tested}
}
\value{
Return ranked data.frame based on "value"
}
\description{
input could be interested gene list. Each row is one gene and columns are "gene" and "value".
"value" could be log2FC or specific value.
colnames of other metadata should be "subtype".
}
